moved to identifyUser
