--- packages/destination-actions/src/destinations/mixpanel/incrementProperty/generated-types.ts
@@ -1,22 +1,0 @@
-// Generated file. DO NOT MODIFY IT BY HAND.
-
-export interface Payload {
-  /**
-   * The IP address of the user. This is only used for geolocation and won't be stored.
-   */
-  ip?: string
-  /**
-   * The unique user identifier set by you
-   */
-  user_id?: string | null
-  /**
-   * The generated anonymous ID for the user
-   */
-  anonymous_id?: string | null
-  /**
-   * Profile properties to increment
-   */
-  $add: {
-    [k: string]: unknown
-  }
-}
